remove parseIntoPtrs() methods and FmtParsedPtrs()
the previous approach which required all flags to be either
pointers or values.  now they may be mixed.  i believe this is a more
flexible API -- and simpler without the parseIntoPtrs() methods.

ParseOptions and ParsedFmtOptions now gain an optional runtime 'ptrs'
field where the user may provide pointers to parse into or format.